% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{cellvel}
\alias{cellvel}
\title{Model-based quantification of cell velocity}
\usage{
cellvel(x, control = NULL)
}
\value{
a list
}
\description{
The functions takes a data.frame, x, as its main input. Meanwhile,
the input list control can be used to configure the MCMC procedure
performed by R-package rstan.  The output is a list which contains:
1) f = fit as rstan object; 2) x = processed input; 3) p = plot
with posterior predictive check; 4) s = summary of model parameters
(means, medians, 95% credible intervals, etc.).
}
\details{
The input x must have cell entries as rows and the following columns:
* replicate = id of the biological replicate (e.g. rep1, rep2, rep3, ...)
* batch = id of the experimental batch (e.g. plate X, plat Y, ...)
* group = treatment group (e.g. compound x at dose 10mmol, ...)
* v = numeric cell speed
}
\examples{
data(d, package = "cellvel")
o <- cellvel(x = d)
head(o)
}
