% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{compare_doses}
\alias{compare_doses}
\title{Compare treatment effect at specific doses (cellvel results)}
\usage{
compare_doses(x)
}
\value{
A data.frame with the following data in each row (comparison):
\itemize{
\item treatment_i = name of treatment i
\item treatment_j = name of treatment j
\item contrast = concatenation of treatment_i and treatment_j,
\item dose = compared dose
\item M, L95, H95 = mean, lower bound and upper bound of difference between the
migration effects of treatment_i and treatment_j
\item \eqn{\pi} = probability of differential effects
\item key = concatenation of treatment_i, treatment_j, and dose
}
}
\description{
The functions compares the posterior distributions of the cell migration
effects of each treatment and dose.
}
\details{
At a given dose, it compares pairs of treatments, i and j. For the comparison
we extract the posterior distributions of their cell migration effects at the
given dose, and then compute an absolute difference between the posteriors.
The result is another posterior distribution (\eqn{\gamma}), with L95 and H95
as the lower and upper bounds of the 95\% highest density interval of
\eqn{\gamma}.

\eqn{\pi} is estimated as:

\eqn{\pi=2\cdot\max{(\int_{-\infty}^{0} \gamma, \int_{0}^{+\infty} \gamma)}-1}.

For treatment groups, i and j, where \eqn{\gamma<0} and the 95\% HDIs of
\eqn{\gamma} lie mostly or completely below 0 (0 = null effect), we have
strong evidence of lower cell migration effect in treatment i compared to
j at the given dose. On the other hand, for treatment groups, i and j,
where \eqn{\gamma} and the 95\% HDI of \eqn{\gamma} lie mostly or completely
above 0, we have strong evidence of higher cell migration effect in treatment
i compared to j. Distributions with the 95\% HDIs more or less centered around
0 indicate that there is no evidence for a clear differnce in the cell migration
effects between the treatments. Note that unclear evidence is not equivalent
to no change, because for a treatment group with \eqn{\gamma\approx 0} we may
also have a wide 95\% HDI, including possibilities for positive or negative
change.
}
\examples{
data(d, package = "compare_doses")
o <- get_cellvel(x = d,
             control = list(mcmc_warmup = 300,
                            mcmc_steps = 600,
                            mcmc_chains = 3,
                            mcmc_cores = 1,
                            mcmc_algorithm = "NUTS",
                            adapt_delta = 0.9,
                            max_treedepth = 10))

u <- compare_doses(x = o)
head(u)
}
