\name{elk}
\alias{elk}
\title{GPS Monitoring of Twenty-three Elk in Canada's Rocky Mountains}
\description{ This dataset contains GPS relocations of twenty-three elk (Cervus elaphus) in Canada's' Rocky Mountains.}
\usage{
data(elk)
}
\format{
This dataset is an object of class \code{\link{ltraj}}.
}
\details{
For convenience, the original dataset was subset to 1 location/day.
}
\source{
Hebblewhite, M. & Merrill, E.H. (2016). Data from: Behavioral flexibility in migratory behavior in a long-lived large herbivore. Movebank Data Repository. URL doi:10.5441/001/1.k8s2g5v7
}

\examples{
  data(elk)
  elk
}
