\name{get_groups}
\alias{get_groups}
\title{Extract group labels}
\description{
  This function extracts treatment group metadata from an object generated by
  \code{cellmig}.
}
\usage{
  get_groups(x)
}
\arguments{
  \item{x}{An object generated by \code{cellmig}, containing treatment group
  names and IDs, compound names and dose.}
}
\details{
  The function retrieves group metadata such as group identifiers, names, 
  compound information, and dose levels.
}
\value{
  A data frame containing the following columns:
  \item{group_id}{Unique identifier for each treatment group.}
  \item{group}{Name of the treatment group.}
  \item{compound}{Compound associated with the treatment group.}
  \item{dose}{Dose level of the compound.}
}
\examples{
data(d, package = "cellmig")
#cellmig_out <- cellmig(x = d,
#           control = list(mcmc_warmup = 300,
#                            mcmc_steps = 1500,
#                            mcmc_chains = 3,
#                            mcmc_cores = 3,
#                            mcmc_algorithm = "NUTS",
#                            adapt_delta = 0.8,
#                            max_treedepth = 10))
data(cellmig_out, package = "cellmig")
u <- get_groups(x = cellmig_out)
}
\seealso{
  \code{cellmig}, \code{get_pairs}, \code{get_groups}, \code{get_violins}
}
