% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputs.R
\name{process_input}
\alias{process_input}
\title{input checks}
\usage{
process_input(x)
}
\value{
a data.frame. missing velocities are removed
}
\description{
The functions takes a data.frame as its main input. It then checks if
the necessary columns are included and processes the input for modeling.
}
\examples{
data(d, package = "cellvel")
pd <- process_input(x = d)
head(pd)
}
