% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{compare_doses}
\alias{compare_doses}
\title{Compare compound effect at specific doses (cellvel results)}
\usage{
compare_doses(x, select_ds, select_ts)
}
\value{
A data.frame with the following data in each row (comparison):
\itemize{
\item compound_i = name of compound i
\item compound_j = name of compound j
\item contrast = concatenation of compound_i and compound_j,
\item dose = compared dose
\item M, L95, H95 = mean, lower bound and upper bound of difference between the
migration effects of compound_i and compound_j
\item \eqn{\pi} = probability of differential effects
\item key = concatenation of compound_i, compound_j, and dose
}
}
\description{
The function \code{compare_doses} compares the posterior distributions of
the cell migration effects of each compound and dose.
}
\details{
The main input, \code{x}, of \code{compare_doses} is the output object from
the function \code{cellvel}. Furthermore, the user can select a subset of
the compounds (input \code{select_ts}, and doses (input \code{select_ds}).

At a given dose, \code{compare_doses} compares pairs of compounds, i and j.
For the comparison we extract the posterior distributions of their cell
migration effects at the given dose, and then compute an absolute difference
between the posteriors. The result is another posterior distribution
(\eqn{\gamma}), with L95 and H95 as the lower and upper bounds of the 95\%
highest density interval of \eqn{\gamma}.

\eqn{\pi} is estimated as: \eqn{\pi=2\cdot\max{(\int_{-\infty}^{0} \gamma,
\int_{0}^{+\infty} \gamma)}-1}.

For compound groups, i and j, where \eqn{\gamma<0} and the 95\% HDIs of
\eqn{\gamma} lie mostly or completely below 0 (0 = null effect), we have
strong evidence of lower cell migration effect in compound i compared to
j at the given dose. On the other hand, for compound groups, i and j, where
\eqn{\gamma} and the 95\% HDI of \eqn{\gamma} lie mostly or completely above 0,
we have strong evidence of higher cell migration effect in compound i compared
to j. Distributions with the 95\% HDIs more or less centered around 0 indicate
that there is no evidence for a clear differnce in the cell migration effects
between the compounds. Note that unclear evidence is not equivalent to no
change, because for a compound group with \eqn{\gamma\approx 0} we may also
have a wide 95\% HDI, including possibilities for positive or negative change.
}
\examples{
data(d, package = "compare_doses")
o <- get_cellvel(x = d,
             control = list(mcmc_warmup = 300,
                            mcmc_steps = 600,
                            mcmc_chains = 3,
                            mcmc_cores = 1,
                            mcmc_algorithm = "NUTS",
                            adapt_delta = 0.9,
                            max_treedepth = 10))

u <- compare_doses(x = o)
head(u)
}
